Rename Postgres tests to shorter name.
