+++ src/test/java/org/blackdread/sqltojava/test/db/postgres/Postgres09TestTest.java
@@ -6,11 +6,11 @@ import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.utility.DockerImageName;
 
-class Postgres15SqlToJdlTransactionPerTestTest extends SqlToJdlTransactionPerTestTest {
+class Postgres09TestTest extends SqlToJdlTransactionPerTestTest {
+    static String dockerImageName;
+
     @Container
-    private static final PostgreSQLContainer POSTGRE_SQL_CONTAINER = new PostgreSQLContainer(
-        DockerImageName.parse("postgres:15beta4-alpine")
-    );
+    private static final PostgreSQLContainer POSTGRE_SQL_CONTAINER = new PostgreSQLContainer(DockerImageName.parse("postgres:9"));
 
     @BeforeAll
     public static void setup() {
